+++ apps/hyperdrive-trading/src/ui/hyperdrive/longs/hooks/useOpenLong.tsx
@@ -4,9 +4,9 @@ import { MutationStatus } from "@tanstack/query-core";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import toast from "react-hot-toast";
 import TransactionToast from "src/ui/base/components/Toaster/TransactionToast";
-import { callWarpcastToast } from "src/ui/base/components/Toaster/WarpcastToast";
 import { SUCCESS_TOAST_DURATION } from "src/ui/base/toasts";
 import { useReadWriteHyperdriveModel } from "src/ui/hyperdrive/hooks/model/useReadWriteHyperdriveModel";
+import { toastWarpcast } from "src/ui/social/components/WarpcastToast";
 import { Address, Hash } from "viem";
 import { usePublicClient } from "wagmi";
 
@@ -71,7 +71,7 @@ export function useOpenLong({
           { id: hash, duration: SUCCESS_TOAST_DURATION },
         );
         setTimeout(() => {
-          callWarpcastToast();
+          toastWarpcast();
         }, SUCCESS_TOAST_DURATION);
         onExecuted?.(hash);
       }
