+++ apps/hyperdrive-trading/src/ui/hyperdrive/lp/hooks/useRemoveLiquidity.tsx
@@ -6,9 +6,9 @@ import {
 } from "@tanstack/react-query";
 import toast from "react-hot-toast";
 import TransactionToast from "src/ui/base/components/Toaster/TransactionToast";
-import { callWarpcastToast } from "src/ui/base/components/Toaster/WarpcastToast";
 import { SUCCESS_TOAST_DURATION } from "src/ui/base/toasts";
 import { useReadWriteHyperdriveModel } from "src/ui/hyperdrive/hooks/model/useReadWriteHyperdriveModel";
+import { toastWarpcast } from "src/ui/social/components/WarpcastToast";
 import { Address, Hash } from "viem";
 import { usePublicClient } from "wagmi";
 
@@ -60,7 +60,7 @@ export function useRemoveLiquidity({
             { id: txHash, duration: SUCCESS_TOAST_DURATION },
           );
           setTimeout(() => {
-            callWarpcastToast();
+            toastWarpcast();
           }, SUCCESS_TOAST_DURATION);
           onExecuted?.(txHash);
         }
